Don't fail if file not found in fallback handler
If the blob storage extension can not find a file remotely, it falls
back to the default local storage backend. If for some reason this
backend can't find the file either (eg because of a corrupted resource
that changed from link to upload) it fails with an IO exception. This
changes it to emit a 404 error instead.