Support for Content-Disposition: inline
Actually all these changes do is supporting an incoming `?preview=1` query
argument on download URLs and forward the `x-disposition = inline` extra to
Giftless when requesting the download URL.

At that point https://github.com/datopian/giftless/pull/90 will kick in
and the storage backend will return the appropiate headers.

The only way to pass this parameter all the way down to the action that
requests the downloadURL was to add a new argument to the download
handlers. This would break backwards compatibility so I've added code to
support plugin hooks that don't support the new `inline` arg.